move T&C to Content component to access queried profile object
